kind of there

+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/model/UIMessage.kt
@@ -152,7 +152,7 @@ sealed class MessageFlowStatus {
 
     object Delivered : MessageFlowStatus()
 
-    data class Read(val count: Int) : MessageFlowStatus()
+    data class Read(val count: Long) : MessageFlowStatus()
 }
 
 data class MessageStatus(
